refactor: entry -> teamへディレクトリ名を変更
