+++ packages/kcms/src/match/adaptor/dummy/mainMatchRepository.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach } from 'vitest';
 import { DummyMainMatchRepository } from './mainMatchRepository.js';
 import { TestRankingMainMatchData } from '../../../testData/match.js';
 import { RunResult, RunResultID } from '../../model/runResult.js';
-import { TeamID } from '../../../entry/models/team.js';
+import { TeamID } from '../../../team/models/team.js';
 import { Option, Result } from '@mikuroxina/mini-fn';
 import { MainMatch, MainMatchID } from '../../model/main.js';
