+++ packages/kcms/src/match/adaptor/dummy/preMatchRepository.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach } from 'vitest';
 import { TestRankingPreMatchData } from '../../../testData/match.js';
 import { PreMatch, PreMatchID } from '../../model/pre.js';
 import { RunResult, RunResultID } from '../../model/runResult.js';
-import { TeamID } from '../../../entry/models/team.js';
+import { TeamID } from '../../../team/models/team.js';
 import { Option, Result } from '@mikuroxina/mini-fn';
 import { DummyPreMatchRepository } from './preMatchRepository.js';
